Import: extract ImporterMmebers
